Update to current Tridactyl repo

+++ native/install.sh
@@ -3,9 +3,9 @@
 set -e
 
 echoerr() {
-    red="\033[31m"
-    normal="\e[0m"
-    echo -e "$red$@$normal" >&2
+    red="\\033[31m"
+    normal="\\e[0m"
+    echo -e "$red$*$normal" >&2
 }
 
 sedEscape() {
@@ -18,8 +18,10 @@ trap "echoerr 'Failed to install!'" ERR
 
 XDG_CONFIG_HOME="${XDG_CONFIG_HOME:-$HOME/.config}/tridactyl"
 XDG_DATA_HOME="${XDG_DATA_HOME:-$HOME/.local/share}/tridactyl"
-manifest_loc="https://raw.githubusercontent.com/tridactyl/tridactyl/master/native/tridactyl.json"
-native_loc="https://raw.githubusercontent.com/tridactyl/tridactyl/master/native/native_main.py"
+
+# Use argument as version or 1.15.0, as that was the last version before we switched to using tags
+manifest_loc="https://raw.githubusercontent.com/tridactyl/tridactyl/${1:-1.15.0}/native/tridactyl.json"
+native_loc="https://raw.githubusercontent.com/tridactyl/tridactyl/${1:-1.15.0}/native/native_main.py"
 
 # Decide where to put the manifest based on OS
 case "$OSTYPE" in
@@ -54,11 +56,21 @@ else
     curl -sS --create-dirs -o "$native_file" "$native_loc"
 fi
 
+if [[ ! -f "$manifest_file" ]] ; then
+    echoerr "Failed to create '$manifest_file'. Please make sure that the directories exist and that you have the necessary permissions."
+    exit 1
+fi
+
+if [[ ! -f "$native_file" ]] ; then
+    echoerr "Failed to create '$native_file'. Please make sure that the directories exist and that you have the necessary permissions."
+    exit 1
+fi
+
 sed -i.bak "s/REPLACE_ME_WITH_SED/$(sedEscape "$native_file_final")/" "$manifest_file"
-chmod +x $native_file
+chmod +x "$native_file"
 
 # Requirements for native messenger
-python_path=$(which python3) || python_path=""
+python_path=$(command -v python3) || python_path=""
 if [[ -x "$python_path" ]]; then
     sed -i.bak "1s/.*/#!$(sedEscape /usr/bin/env) $(sedEscape "$python_path")/" "$native_file"
     mv "$native_file" "$native_file_final"
